+ odpowiednia zmiana metod, aby JPARepisotry działało poprawnie
